backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pdfplumber
pillow
openai>=1.32.0
